refactor : Hr 관련 컴포넌트 파일 위치 이동
HrDetail 폴더 생성 후 아래로 파일 위치 이동
