+++ Front-End/src/components/HRManagement/HrDetail/HrEdu.js
